_references/FriendsOfGalaxy/update-test-integration/requirements.txt
PyGithub==1.43.8
invoke==1.3.0
